fix problem with transparency + problem with floats

+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: mbueno-g <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/31 17:00:10 by mbueno-g          #+#    #+#             */
-/*   Updated: 2022/02/06 21:20:31 by mbueno-g         ###   ########.fr       */
+/*   Updated: 2022/02/07 16:29:00 by mbueno-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int	cub_atoi(const char *nptr, short *nbr)
 	return (0);
 }
 
-float	degree_to_radians(int degree)
+float	degree_to_radians(float degree)
 {
 	return (degree * M_PI / 180);
 }
